Refactor: moves classes to appropriate files
