+++ lab/AboutDialog/AboutDialog/BindableBase.cs
@@ -13,7 +13,7 @@ namespace Sharpen
         /// <summary>
         /// Multicast event for property change notifications.
         /// </summary>
-        public event PropertyChangedEventHandler PropertyChanged;
+        public event PropertyChangedEventHandler? PropertyChanged;
 
         /// <summary>
         /// Checks if a property already matches a desired value.  Sets the property and
@@ -31,7 +31,7 @@ namespace Sharpen
         /// True if the value was changed, false if the existing value matched the
         /// desired value.
         /// </returns>
-        protected bool SetProperty<T>(ref T storage, T value, [CallerMemberName] string propertyName = null)
+        protected bool SetProperty<T>(ref T storage, T value, [CallerMemberName] string propertyName = "#default#")
         {
             if (Equals(storage, value))
             {
@@ -51,7 +51,7 @@ namespace Sharpen
         /// value is optional and can be provided automatically when invoked from compilers
         /// that support <see cref="CallerMemberNameAttribute" />.
         /// </param>
-        protected void OnPropertyChanged([CallerMemberName] string propertyName = null)
+        protected void OnPropertyChanged([CallerMemberName] string propertyName = "#default#")
         {
             PropertyChanged?.Invoke(this, new PropertyChangedEventArgs(propertyName));
         }
