+++ lab/AboutDialog/AboutDialog/ConfigurationBaseViewModel.cs
@@ -1,0 +1,12 @@
+﻿using System.Configuration;
+
+namespace Sharpen
+{
+    /// <summary>
+    /// Base class for view models that need access to AboutDialogConfigurationSection.
+    /// </summary>
+    internal class ConfigurationBaseViewModel : BindableBase
+    {
+        protected AboutDialogConfigurationSection configuration = (AboutDialogConfigurationSection)ConfigurationManager.GetSection("aboutDialogConfigurationSection");
+    }
+}
